Rewote money amount types
Instead of having a protocol that two separate types conform to, we use a base generic type that can use different coding keys for encoding/decoding the currency and value
We also changed the  type from String to Decimal. This removes the need to validation of the  property and feels more native. The custom encoding/decoding methods use string<->decimal conversion for the value